refactor: Modal 추상 클래스를 만들어서 상속하게 만들다
